DEV-108 refactor: reduce highlight folder complexity
- Resolve `highlights/maps/`` folder into just `highlights/`
- Move init functions from `highlights/init.lua` to `utils/highlights`
