+++ lua/terra-core/highlights/plugins/__plugin_highlight_template.lua
