+++ lua/terra-core/highlights/plugins/gitgutter.lua
