+++ lua/terra-core/highlights/plugins/hop.lua
