+++ lua/terra-core/highlights/plugins/neotree.lua
