+++ lua/terra-core/highlights/plugins/nvim_tree.lua
