+++ lua/terra-core/highlights/plugins/telescope.lua
