+++ lua/terra-core/highlights/plugins/ts_rainbow.lua
